meeting_workflow/requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
google-generativeai>=0.3.0
google-adk>=0.5.0 